test(cli): split Claws package fixture
