+++ src/cli/claws-cli.test-helpers.ts
@@ -1,4 +1,4 @@
-import { realpath, writeFile } from "node:fs/promises";
+import { mkdir, realpath, writeFile } from "node:fs/promises";
 import { basename, dirname, join } from "node:path";
 
 export const minimalManifest = {
@@ -31,4 +31,41 @@ export async function writeManifestFile(
   const path = join(dir, "openclaw.claw.json");
   await writeFile(path, JSON.stringify(value), "utf8");
   return path;
+}
+
+export async function writePackageFixture(tempDirs: {
+  make(prefix: string): string;
+}): Promise<{ root: string; workspace: string }> {
+  const root = tempDirs.make("openclaw-claws-cli-package-");
+  await mkdir(join(root, "workspace"));
+  await writeFile(join(root, "workspace", "AGENTS.md"), "# Demo\n", "utf8");
+  await writeFile(
+    join(root, "package.json"),
+    JSON.stringify({
+      name: "@acme/demo-agent",
+      version: "1.2.3",
+      openclaw: { claw: "openclaw.claw.json" },
+    }),
+    "utf8",
+  );
+  await writeFile(
+    join(root, "openclaw.claw.json"),
+    JSON.stringify({
+      schemaVersion: 1,
+      agent: { id: "demo-agent", name: "Demo Agent" },
+      workspace: {
+        bootstrapFiles: { "AGENTS.md": { source: "workspace/AGENTS.md" } },
+      },
+      packages: [
+        {
+          kind: "skill",
+          source: "clawhub",
+          ref: "@acme/demo-skill",
+          version: "1.0.0",
+        },
+      ],
+    }),
+    "utf8",
+  );
+  return { root, workspace: join(root, "target-workspace") };
 }
