+++ src/cli/claws-cli.test.ts
@@ -120,41 +120,6 @@ async function writeManifest(value: unknown = cliTestHelpers.minimalManifest): P
   return await cliTestHelpers.writeManifestFile(tempDirs, value);
 }
 
-async function writePackage(): Promise<{ root: string; workspace: string }> {
-  const root = tempDirs.make("openclaw-claws-cli-package-");
-  await mkdir(join(root, "workspace"));
-  await writeFile(join(root, "workspace", "AGENTS.md"), "# Demo\n", "utf8");
-  await writeFile(
-    join(root, "package.json"),
-    JSON.stringify({
-      name: "@acme/demo-agent",
-      version: "1.2.3",
-      openclaw: { claw: "openclaw.claw.json" },
-    }),
-    "utf8",
-  );
-  await writeFile(
-    join(root, "openclaw.claw.json"),
-    JSON.stringify({
-      schemaVersion: 1,
-      agent: { id: "demo-agent", name: "Demo Agent" },
-      workspace: {
-        bootstrapFiles: { "AGENTS.md": { source: "workspace/AGENTS.md" } },
-      },
-      packages: [
-        {
-          kind: "skill",
-          source: "clawhub",
-          ref: "@acme/demo-skill",
-          version: "1.0.0",
-        },
-      ],
-    }),
-    "utf8",
-  );
-  return { root, workspace: join(root, "target-workspace") };
-}
-
 async function runCli(args: string[]) {
   const program = new Command();
   program.exitOverride();
@@ -424,7 +389,7 @@ describe("claws cli", () => {
   });
 
   it("takes identity from package.json and plans one new agent", async () => {
-    const { root, workspace } = await writePackage();
+    const { root, workspace } = await cliTestHelpers.writePackageFixture(tempDirs);
     const expectedWorkspace = await cliTestHelpers.canonicalFuturePath(workspace);
 
     await runCli(["claws", "add", root, "--dry-run", "--workspace", workspace, "--json"]);
@@ -462,7 +427,7 @@ describe("claws cli", () => {
   });
 
   it("blocks adding into an existing agent instead of merging", async () => {
-    const { root, workspace } = await writePackage();
+    const { root, workspace } = await cliTestHelpers.writePackageFixture(tempDirs);
     mocks.loadConfig.mockReturnValue({ agents: { entries: { "demo-agent": {} } } });
 
     await runCli(["claws", "add", root, "--dry-run", "--workspace", workspace, "--json"]);
@@ -475,7 +440,7 @@ describe("claws cli", () => {
   });
 
   it("honors an explicit unused agent id in the plan", async () => {
-    const { root, workspace } = await writePackage();
+    const { root, workspace } = await cliTestHelpers.writePackageFixture(tempDirs);
     mocks.loadConfig.mockReturnValue({ agents: { entries: { "demo-agent": {} } } });
 
     await runCli([
@@ -807,7 +772,7 @@ describe("claws cli", () => {
   });
 
   it("prints a read-only grouped update plan", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
 
     await runCli(["claws", "update", "demo-agent", "--from", root, "--dry-run", "--json"]);
 
@@ -831,7 +796,7 @@ describe("claws cli", () => {
   });
 
   it("prints capability escalation details in human update previews", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
 
     await runCli(["claws", "update", "demo-agent", "--from", root, "--dry-run"]);
 
@@ -848,7 +813,7 @@ describe("claws cli", () => {
   });
 
   it("returns failure when an update plan contains blocked actions", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
     mocks.buildClawUpdatePlan.mockResolvedValueOnce({
       schemaVersion: "openclaw.clawUpdatePlan.v1",
       stability: "experimental",
@@ -891,7 +856,7 @@ describe("claws cli", () => {
   });
 
   it("uses the source recorded by the installed Claw when --from is omitted", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
     mocks.readClawStatus.mockResolvedValue({
       schemaVersion: "openclaw.clawStatus.v1",
       records: [
@@ -945,7 +910,7 @@ describe("claws cli", () => {
   });
 
   it("fails closed when update is invoked without dry-run", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
 
     await runCli(["claws", "update", "demo-agent", "--from", root, "--json"]);
 
@@ -958,7 +923,7 @@ describe("claws cli", () => {
   });
 
   it("requires exact plan integrity with update consent", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
 
     await runCli(["claws", "update", "demo-agent", "--from", root, "--yes", "--json"]);
 
@@ -970,7 +935,7 @@ describe("claws cli", () => {
   });
 
   it("applies a supported update only after explicit consent", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
 
     await runCli([
       "claws",
@@ -1011,7 +976,7 @@ describe("claws cli", () => {
   });
 
   it("reports uncertain update mutations as partial JSON", async () => {
-    const { root } = await writePackage();
+    const { root } = await cliTestHelpers.writePackageFixture(tempDirs);
     mocks.applyClawUpdatePlan.mockRejectedValueOnce(
       new ClawUpdateMutationError("update_partial", "artifact outcome requires reconciliation"),
     );
